+++ src/50expression.js
@@ -238,7 +238,8 @@ yy.Op.prototype.toJavaScript = function(context,tableid,defcols) {
 		if(this.right instanceof yy.Select ) {
 			var s = '(';
 //			s += 'this.query.queriesdata['+this.queriesidx+']';
-			s += 'alasql.utils.flatArray(this.query.queriesfn['+(this.queriesidx)+'](params,null,context))';
+//			s += 'alasql.utils.flatArray(this.query.queriesfn['+(this.queriesidx)+'](params,null,context))';
+			s += 'alasql.utils.flatArray(this.queriesfn['+(this.queriesidx)+'](params,null,context))';
 			s += '.indexOf(';
 			s += this.left.toJavaScript(context,tableid, defcols)+')>-1)';
 			return s;
@@ -262,7 +263,8 @@ yy.Op.prototype.toJavaScript = function(context,tableid,defcols) {
 		if(this.right instanceof yy.Select ) {
 			var s = '('
 				//this.query.queriesdata['+this.queriesidx+']
-			s += 'alasql.utils.flatArray(this.query.queriesfn['+(this.queriesidx)+'](params,null,p))';
+//			s += 'alasql.utils.flatArray(this.query.queriesfn['+(this.queriesidx)+'](params,null,p))';
+			s += 'alasql.utils.flatArray(this.queriesfn['+(this.queriesidx)+'](params,null,p))';
 			s +='.indexOf(';
 			s += this.left.toJavaScript(context,tableid, defcols)+')<0)';
 			return s;