+++ test/test297.js
@@ -18,11 +18,29 @@ describe('Test 297 INSERT,DELETE,UDPATE with subqueries', function() {
 
   it('2. DELETE',function(done){
     var res = alasql('DELETE FROM one WHERE a = (SELECT MAX(a) FROM one)');
-    console.log(res);
     assert.deepEqual(res,1);
     done();
   });
 
+  it('3. UPDATE',function(done){
+    var res = alasql('UPDATE one SET b = 100 WHERE a = (SELECT MAX(a) FROM one)');
+    assert.deepEqual(res,1);
+    var res = alasql('SELECT * FROM one');
+    assert.deepEqual(res,
+        [ { a: 1, b: 10 }, { a: 2, b: 100 } ]
+    );
+    done();
+  });
+
+  it('4. INSERT',function(done){
+    var res = alasql('INSERT INTO one VALUES (5,(SELECT MAX(b) FROM one)+1)');
+    assert.deepEqual(res,1);
+    var res = alasql('SELECT * FROM one');
+//    console.log(res);
+    assert.deepEqual(res,
+        [ { a: 1, b: 10 }, { a: 2, b: 100 }, { a: 5, b: 101 } ]);
+    done();
+  });
 
   it('99. DROP DATABASE',function(done){
     alasql('DROP DATABASE test297');