+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
+
 ### 0.1.6 "Palermo" (13.05.2015 - ...05.2015)
 * SET NOCOUNT OFF (for CREATE and INSERT)
+* ROWNUM() and ROW_NUMBER() functions
 *
 
 ### 0.1.5 "San Marino" (12.05.2015 - 12.05.2015)