+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-### 0.1.6 "Palermo" (13.05.2015 - ...05.2015)
+### 0.1.6 "Palermo" (13.05.2015 - 15.05.2015)
 * SET NOCOUNT OFF (for CREATE and INSERT)
 * ROWNUM() and ROW_NUMBER() functions
 * Promised version of alasql() - alasql.async() (based on es6-promises)
 * SELECT * FROM Json
 * SEARCH COMMA selector
 * Fixed bug with ORDER BY 1,2,3
+* Added subqueries for INSERT/DELETE/UPDATE
 
 ### 0.1.5 "San Marino" (12.05.2015 - 12.05.2015)
 * Added Meteor package (agershun:alasql) - still does not work - skeleton